+++ packages/@expo/cli/src/api/user/actions.ts
@@ -31,7 +31,9 @@ export async function showLoginPromptAsync({
   }
   const hasCredentials = options.username && options.password;
   const sso = options.sso;
-  const browser = options.browser;
+
+  // Browser-based login is the default.
+  const browser = options.browser ?? (!hasCredentials && isInteractive());
 
   if (printNewLine) {
     Log.log();
@@ -39,6 +41,7 @@ export async function showLoginPromptAsync({
 
   if (sso || browser) {
     await browserLoginAsync({ sso: !!sso });
+    Log.log('Logged in');
     return;
   }
 
@@ -94,6 +97,8 @@ export async function showLoginPromptAsync({
       throw e;
     }
   }
+
+  Log.log('Logged in');
 }
 
 export async function tryGetUserAsync(): Promise<Actor | null> {